requirements.txt
flask
pyJWT
argon2-cffi
Flask-SQLAlchemy